feat: added examples of typography with Bootstrap
Added HTML tags and Bootstrap classes to display various typography elements. The h1 to h6 tags are used for headings of different sizes, the p tag is used to highlight text, and the ul and ol tags are used for unordered and ordered lists, respectively. Bootstrap classes are used to apply formatting styles.