feat: added align-items property in Flexbox with Bootstrap
Implemented Flexbox layouts using the `align-items` property in Bootstrap.
The code includes multiple container divs with different `align-items` values.
Each container div contains three child divs with background colors and content.
The `align-items-stretch` class is used to stretch the child divs to match the container's height, `align-items-start` to align them to the start, `align-items-center` to center them vertically, and `align-items-end` to align them to the end.
Additionally, a combination of `align-items-end` and `justify-content-center` classes is used to align the child divs to the end vertically and center them horizontally.
Lastly, there is an `align-items-lg-center` class used to align the child divs to the center vertically on larger screens.

Refs: https://youtu.be/NgBk_-7k3hw