feat: added Bootstrap Carousel component
Implemented Bootstrap Carousel component with a slideshow of images.

The carousel is configured with three slides that display images in a responsive manner.

It includes previous and next navigation buttons using the `carousel-control-prev` and `carousel-control-next` classes.

The carousel is set to automatically cycle through the slides.

Images used in the carousel: image_1.jpeg, image_2.jpeg, image_3.jpeg.

Refs: https://youtu.be/ZnqNp9VCtko